refactored code moved pages to separate folder home page currently search page
